update organisation image global hook
